+++ renderer/Model_ase.cpp
@@ -781,7 +781,7 @@ aseModel_t *ASE_Parse( const char *buffer, bool verbose ) {
 	ase.verbose = verbose;
 
 	ase.buffer = buffer;
-	ase.len = strlen( buffer );
+    ase.len = static_cast<int>(strlen(buffer));
 	ase.curpos = ase.buffer;
 	ase.currentObject = NULL;
