+++ renderer/Model_md3.cpp
@@ -150,7 +150,7 @@ void idRenderModelMD3::InitFromFile( const char *fileName ) {
 
 		// strip off a trailing _1 or _2
 		// this is a crutch for q3data being a mess
-		j = strlen( surf->name );
+        j = static_cast<int>(strlen(surf->name));
 		if ( j > 2 && surf->name[j-2] == '_' ) {
 			surf->name[j-2] = 0;
 		}
